# performance-only fixes (np.clip, np.array growing, )