--- a/causal_world/envs/robot/trifinger.py
+++ b/causal_world/envs/robot/trifinger.py
@@ -148,14 +148,6 @@ class TriFingerRobot(object):
             [joint[1] for joint in current_joint_states])
         current_torques = np.array([joint[3] for joint in current_joint_states])
         
-        a_current_joint_states = np.array(current_joint_states)
-        a_current_position = a_current_joint_states[:,0]
-        a_current_velocity = a_current_joint_states[:,1]
-        a_current_torques = a_current_joint_states[:,3]
-        np.testing.assert_array_equal(current_position, a_current_position)
-        np.testing.assert_array_equal(current_velocity, a_current_velocity)
-        np.testing.assert_array_equal(current_torques, a_current_torques)
-        
         self._latest_full_state = {
             'positions':
                 current_position,
@@ -424,11 +416,13 @@ class TriFingerRobot(object):
     
     def _safety_torque_check(self, desired_torques):
         """
-
+        limits desired_torques to max_motor_torque 
+        and lowers amplitude of desired_torques in velocity direction
+        
         :param desired_torques: (nd.array) the desired torque commands to be applied
                                             to the robot.
 
-        :return: (nd.array) the modified torque commands after applying a safety check.
+        :return: (list) the modified torque commands after applying a safety check.
         """       
         applied_torques = clip(
             np.asarray(desired_torques),
@@ -436,14 +430,14 @@ class TriFingerRobot(object):
             self._max_motor_torque,
         )
 
-        applied_torques = list(clip(
-            applied_torques \
-            - self._safety_kd * self._latest_full_state['velocities'],
-            -self._max_motor_torque,
-            self._max_motor_torque,
-        ))
+        applied_torques = clip(
+                applied_torques \
+                - self._safety_kd * self._latest_full_state['velocities'],
+                -self._max_motor_torque,
+                self._max_motor_torque,
+        )
         
-        return applied_torques
+        return list(applied_torques)
 
     def inverse_kinematics(self, desired_tip_positions, rest_pose):
         """
@@ -608,89 +602,44 @@ class TriFingerRobot(object):
 
         :return: (nd.array) the current end effector positions of the robot.
         """
-        # 37.4ms
-        result = np.array([])
         if self._pybullet_client_full_id is not None:
             position_1 = pybullet.getLinkState(
                 WorldConstants.ROBOT_ID,
                 linkIndex=5,
                 computeForwardKinematics=True,
-                physicsClientId=self._pybullet_client_full_id)
+                physicsClientId=self._pybullet_client_full_id)[0]
             position_2 = pybullet.getLinkState(
                 WorldConstants.ROBOT_ID,
                 linkIndex=10,
                 computeForwardKinematics=True,
-                physicsClientId=self._pybullet_client_full_id)
+                physicsClientId=self._pybullet_client_full_id)[0]
             position_3 = pybullet.getLinkState(
                 WorldConstants.ROBOT_ID,
                 linkIndex=15,
                 computeForwardKinematics=True,
-                physicsClientId=self._pybullet_client_full_id)
+                physicsClientId=self._pybullet_client_full_id)[0]
         else:
             position_1 = pybullet.getLinkState(
                 WorldConstants.ROBOT_ID,
                 linkIndex=5,
                 computeForwardKinematics=True,
-                physicsClientId=self._pybullet_client_w_o_goal_id)
+                physicsClientId=self._pybullet_client_w_o_goal_id)[0]
             position_2 = pybullet.getLinkState(
                 WorldConstants.ROBOT_ID,
                 linkIndex=10,
                 computeForwardKinematics=True,
-                physicsClientId=self._pybullet_client_w_o_goal_id)
+                physicsClientId=self._pybullet_client_w_o_goal_id)[0]
             position_3 = pybullet.getLinkState(
                 WorldConstants.ROBOT_ID,
                 linkIndex=15,
                 computeForwardKinematics=True,
-                physicsClientId=self._pybullet_client_w_o_goal_id)
-        result = np.append(result, position_1[0])
-        result = np.append(result, position_2[0])
-        result = np.append(result, position_3[0])
+                physicsClientId=self._pybullet_client_w_o_goal_id)[0]
+        result = list(position_1) + list(position_2) + list(position_3)
         result[2] -= WorldConstants.FLOOR_HEIGHT
         result[5] -= WorldConstants.FLOOR_HEIGHT
         result[-1] -= WorldConstants.FLOOR_HEIGHT
-    
-        # 23ms
-        if self._pybullet_client_full_id is not None:
-            position_1 = pybullet.getLinkState(
-                WorldConstants.ROBOT_ID,
-                linkIndex=5,
-                computeForwardKinematics=True,
-                physicsClientId=self._pybullet_client_full_id)[0]
-            position_2 = pybullet.getLinkState(
-                WorldConstants.ROBOT_ID,
-                linkIndex=10,
-                computeForwardKinematics=True,
-                physicsClientId=self._pybullet_client_full_id)[0]
-            position_3 = pybullet.getLinkState(
-                WorldConstants.ROBOT_ID,
-                linkIndex=15,
-                computeForwardKinematics=True,
-                physicsClientId=self._pybullet_client_full_id)[0]
-        else:
-            position_1 = pybullet.getLinkState(
-                WorldConstants.ROBOT_ID,
-                linkIndex=5,
-                computeForwardKinematics=True,
-                physicsClientId=self._pybullet_client_w_o_goal_id)[0]
-            position_2 = pybullet.getLinkState(
-                WorldConstants.ROBOT_ID,
-                linkIndex=10,
-                computeForwardKinematics=True,
-                physicsClientId=self._pybullet_client_w_o_goal_id)[0]
-            position_3 = pybullet.getLinkState(
-                WorldConstants.ROBOT_ID,
-                linkIndex=15,
-                computeForwardKinematics=True,
-                physicsClientId=self._pybullet_client_w_o_goal_id)[0]
-        result3 = list(position_1) + list(position_2) + list(position_3)
-        result3[2] -= WorldConstants.FLOOR_HEIGHT
-        result3[5] -= WorldConstants.FLOOR_HEIGHT
-        result3[-1] -= WorldConstants.FLOOR_HEIGHT
-        result3 = np.array(result3)
-
-        np.testing.assert_array_equal(result,result3)
         
-        return result
+        return np.array(result)
     
     def _process_action_joint_positions(self, robot_state):
         """