as discussed in #101:
- `np.clip` is slower than `np.core.umath.clip` (saves ~10%) simulation time
- lists instead of np.array growing and smaller fixes (saves ~10% simulation time)

Time before for 10 runs for pusing task (200 steps,3 resets) : `~13.0s +- 200ms `after `10.8s +- 150ms`

All changed functions have been like: 
`np.testing.assert_array_equal(prior_implementation, faster_implementation)` while running the pytests